StackMachine: move the implementation into the header
Necessary as fhe_cc_library accepts a single header file.  We want to
experiment with two FHE libraries, one where StackMachine::tick() is
hls_top.  Since we can't have both methods be marked this way, we need
to include the declaration and definition in a single header file, which
is then included into the top-level .cc file for the fhe library, with a
function marked hls_top that includes the respective method.  This will
be added in a subsequent commit.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>