checkpoint: wip
The stack_machine_tick_interpreted_tfhe target builds with a 50-command
buffer, 50-slot stack, and 10-deep callstack.

The stack_machine_tick_interpreted_tfhe_testbench is WIP: does not build
yet, I need to rework the API to take in a cost StackMachine and return
a StackMachine as the result, rather than update in-place.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>