Fix handling of return values as tuples in CollectOutputs
The return value can be a tuple in two cases:

Case A:
   StructA foo(StructB in)
   void foo(Struct &in)
Case B:
   StructR foo(StructA a, StructB &b)
   StructR foo(StructA &a, StructB b)
   void foo(StructA &a, StructB &b)

IOW, Case A is when a function returns a single struct (either explicitly or as
an non-const reference, this struct is represented as a tuple in the return
type.  CaseB is when multiple values are returned (via some combination of a
return value and one or more non-const references, or a void return and two or
more non-const references.

Thus in case A we want to append to elements, while in case B we want to splice
the results tuple in.

In both cases, type->kind() will be a kTuple, so we need an additional check to
tell cases A and B apart.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>